// src/types/banco.types.ts
export interface Banco {
    idBancos?: number;
    IdBancos?: number;
    nombre?: string;
    Nombre?: string;
    ruc?: string;
    Ruc?: string;
    direccion?: string;
    Direccion?: string;
    telefonos?: string;
    estado?: string | boolean;
    Estado?: string | boolean;
    swift?: string;
}